+++ src/main/java/nl/prbed/hu/aviation/presentation/dto/UpdateAirportResponseDto.java
@@ -8,7 +8,7 @@ import nl.prbed.hu.aviation.domain.City;
 @RequiredArgsConstructor
 public class UpdateAirportResponseDto {
     private final String code;
-    private final float latitude;
-    private final float longitude;
+    private final double latitude;
+    private final double longitude;
     private final City city;
 }
